refactor(solver): split inference candidate classification out of inference_helpers.rs
`inference_helpers.rs` was 2065 lines, exceeding the 2000-line limit.
Extract the five candidate-classification predicates
(`is_mergeable_direct_inference_candidate`,
`inference_type_contains_fresh_object_or_array`,
`is_structural_return_inference_candidate`,
`has_conflicting_literal_bases`, `primitive_base_of`) into a new
focused module `inference_helpers_candidates.rs` (173 lines).

Three previously-private methods get `pub(super)` so they remain
callable from `inference_helpers.rs` across the module boundary.
No logic change; all 6611 passing solver tests continue to pass.
